Refactor representor instantiating.
Centralized a bunch of the 'representor'-related functionality to a
single class. This makes the Resource and Ketting classes a tad easier
to understand.

It's not quite simple enough yet, but it's making headway.
